+++ applications/scene-display/examples/character.xml
@@ -176,7 +176,7 @@
     </light>
     <!-- Sky -->
     <sky id="sky" camera="main-camera" update-interval="200.0" >
-        <star-map texture="res/sky/milkyway.png" scattering="0.8" delta-magnitude="0.5" lod="0" />
+        <star-map texture="res/sky/milkyway.png" scattering="0.8" delta-magnitude="0.05" lod="0" />
         <stars catalog="res/sky/bright-stars.bin" scale="1.4"
             scattering="1.0" scintillation="0.2" apparent-magnitude="6.0"
             glare-intensity="0.04" glare-scale="0.2"
