--- regen/external/osghimmel/astronomy.h
@@ -1,69 +1,0 @@
-// Copyright (c) 2011-2012, Daniel Müller <dm@g4t3.de>
-// Computer Graphics Systems Group at the Hasso-Plattner-Institute, Germany
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without 
-// modification, are permitted provided that the following conditions are met:
-//   * Redistributions of source code must retain the above copyright notice, 
-//     this list of conditions and the following disclaimer.
-//   * Redistributions in binary form must reproduce the above copyright 
-//     notice, this list of conditions and the following disclaimer in the 
-//     documentation and/or other materials provided with the distribution.
-//   * Neither the name of the Computer Graphics Systems Group at the 
-//     Hasso-Plattner-Institute (HPI), Germany nor the names of its 
-//     contributors may be used to endorse or promote products derived from 
-//     this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" 
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE 
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE 
-// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE 
-// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR 
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF 
-// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS 
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN 
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE 
-// POSSIBILITY OF SUCH DAMAGE.
-
-#pragma once
-#ifndef OSG_HIMMEL_ASTRONOMY_H_
-#define OSG_HIMMEL_ASTRONOMY_H_
-
-#include "declspec.h"
-#include "abstractastronomy.h"
-
-namespace osgHimmel {
-	class Astronomy : public AbstractAstronomy {
-	public:
-		Astronomy();
-
-	protected:
-		regen::Vec3f moonPosition(
-				const t_aTime &aTime, float latitude, float longitude,
-				bool refractionCorrected) const final;
-
-		regen::Vec3f sunPosition(
-				const t_aTime &aTime, float latitude, float longitude,
-				bool refractionCorrected) const final;
-
-		regen::Mat4f moonOrientation(const t_aTime &aTime, float latitude, float longitude) const final;
-
-		float earthShineIntensity(const t_aTime &aTime, float latitude, float longitude) const final;
-
-		regen::Mat4f equToHorTransform(const t_aTime &aTime, float latitude, float longitude) const final;
-
-		float sunDistance(const t_julianDay &t) const final;
-
-		float angularSunRadius(const t_julianDay &t) const final;
-
-		float moonRadius() const final;
-
-		float moonDistance(const t_julianDay &t) const final;
-
-		float angularMoonRadius(const t_julianDay &t) const final;
-	};
-
-} // namespace osgHimmel
-
-#endif // OSG_HIMMEL_ASTRONOMY_H_
