+++ regen/sky/bright-stars.cpp
@@ -1,13 +1,12 @@
-#include "stars.h"
+#include "bright-stars.h"
 
 #include <regen/textures/texture-loader.h>
-#include <regen/external/osghimmel/brightstars.h>
 #include <regen/external/osghimmel/randommapgenerator.h>
 #include <regen/external/osghimmel/coords.h>
 
 using namespace regen;
 
-Stars::Stars(const ref_ptr<Sky> &sky)
+BrightStars::BrightStars(const ref_ptr<Sky> &sky)
 		: SkyLayer(sky) {
 	state()->joinStates(ref_ptr<BlendState>::alloc(GL_SRC_ALPHA, GL_ONE));
 
@@ -61,19 +60,54 @@ Stars::Stars(const ref_ptr<Sky> &sky)
 #define _rightasc(deg, min, sec) \
     (_rad(_rightascd(deg, min, sec)))
 
-void Stars::set_brightStarsFile(const std::string &brightStars) {
-	osgHimmel::BrightStars bs(brightStars.c_str());
-	const osgHimmel::BrightStars::s_BrightStar *stars = bs.stars();
-	if (bs.numStars() == 0) {
+// Star model is based on the extended bright star catalogue.
+struct s_BrightStar {
+	float Vmag;   // visual magnitude (mag)
+	float RA;     // right ascension (decimal hours)
+	float DE;     // declination (decimal degrees)
+	float pmRA;   // proper annual motion in right ascension (decimal hours)
+	float pmDE;   // proper annual motion in declination (decimal degrees)
+	float sRGB_R; // approximated color, red value   ]0;1[
+	float sRGB_G; // approximated color, green value ]0;1[
+	float sRGB_B; // approximated color, blue value  ]0;1[
+};
+
+static std::vector<s_BrightStar> loadStarsData(const char *fileName) {
+	// Retrieve file size.
+	FILE *f;
+#ifdef __GNUC__
+	f = std::fopen(fileName, "r");
+#else // __GNUC__
+	fopen_s(&f, fileName, "r");
+#endif // __GNUC__
+	if (!f) { return {}; }
+
+	std::fseek(f, 0, SEEK_END);
+	const auto fileSize = std::ftell(f);
+	std::fclose(f);
+
+	auto numStars = fileSize / sizeof(s_BrightStar);
+
+	std::vector<s_BrightStar> stars(numStars);
+	std::ifstream in_stream(fileName, std::ios::binary);
+	in_stream.read(reinterpret_cast<char *>(stars.data()), fileSize);
+	return stars;
+}
+
+void BrightStars::set_brightStarsFile(const std::string &brightStars) {
+	auto starsData = loadStarsData(brightStars.c_str());
+	numStars_ = static_cast<unsigned int>(starsData.size());
+	if (numStars_ == 0) {
 		REGEN_WARN("Unable to load bright stars catalog at " << brightStars << ".");
 		return;
 	}
-	REGEN_INFO("Loaded " << bs.numStars() << " bright stars from " << brightStars << ".");
+	REGEN_INFO("Loaded " << numStars_ << " bright stars from " << brightStars << ".");
 
-	pos_->setVertexData(bs.numStars());
-	col_->setVertexData(bs.numStars());
+	pos_->setVertexData(numStars_);
+	col_->setVertexData(numStars_);
 
-	for (unsigned int i = 0; i < bs.numStars(); ++i) {
+	auto stars = starsData.data();
+	for (unsigned int i = 0; i < numStars_; ++i) {
 		osgHimmel::t_equf equ;
 		equ.right_ascension = _rightascd(stars[i].RA, 0, 0);
 		equ.declination = stars[i].DE;
@@ -93,7 +127,7 @@ void Stars::set_brightStarsFile(const std::string &brightStars) {
 	meshState_->end();
 }
 
-void Stars::updateNoiseTexture() {
+void BrightStars::updateNoiseTexture() {
 	const int noiseN = 256;
 
 	byte *noiseMap = new byte[noiseN];
@@ -118,14 +152,14 @@ void Stars::updateNoiseTexture() {
 	noiseTexState_->set_name("noiseTexture");
 }
 
-float Stars::defaultApparentMagnitude() { return 7.0f; }
+float BrightStars::defaultApparentMagnitude() { return 7.0f; }
 
-Vec3f Stars::defaultColor() { return {0.66, 0.78, 1.0}; }
+Vec3f BrightStars::defaultColor() { return {0.66, 0.78, 1.0}; }
 
-float Stars::defaultColorRatio() { return 0.66f; }
+float BrightStars::defaultColorRatio() { return 0.66f; }
 
-float Stars::defaultGlareScale() { return 1.2f; }
+float BrightStars::defaultGlareScale() { return 1.2f; }
 
-float Stars::defaultScintillation() { return 0.2f; }
+float BrightStars::defaultScintillation() { return 0.2f; }
 
-float Stars::defaultScattering() { return 2.0f; }
+float BrightStars::defaultScattering() { return 2.0f; }
